Enhace Button comonent
* move `ButtonSpinner` to `button-spinner.tsx`
* Add comments on `Button` component
